requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.3.2
flask-cors==5.0.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
geographiclib==2.0
geopy==2.3.0
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==1.23.5
pandas==1.5.3
psycopg2-binary==2.9.6
pyodbc==4.0.39
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==1.4.46
typing_extensions==4.13.2
Werkzeug==3.1.3
